backend/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cryptography==44.0.2
cycler==0.12.1
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.17.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.55.8
fsspec==2025.2.0
future==0.18.2
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.7
macholib==1.15.2
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
psycopg2-binary==2.9.9
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.1
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rsa==4.9
six==1.15.0
sniffio==1.3.1
spotipy==2.25.0
SQLAlchemy==2.0.38
starlette==0.46.0
sympy==1.13.1
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.27.0
Werkzeug==3.1.3
zipp==3.21.0
gunicorn==21.2.0
